refactor(extensions): extract extension API type contracts
